wo#7092 . move debug helpers to seam
