+++ tests/unit/libpluto/lp86-h2h-invalid-deleteSA-R2-R/deleteChildSA-invalid-fromR2.c
@@ -20,25 +20,6 @@ static void init_fake_secrets(void)
 static void init_loaded(void)
 {   /* nothing */ }
 
-#define WANT_THIS_DBG DBG_EMITTING|DBG_PARSING|DBG_CONTROL|DBG_CONTROLMORE|DBG_CRYPT|DBG_PRIVATE
-
-void enable_debugging(void)
-{
-    base_debugging = WANT_THIS_DBG;
-    reset_debugging();
-}
-
-void enable_debugging_on_sa(int num)
-{
-    struct state *st;
-    lset_t to_enable = WANT_THIS_DBG;
-    st = state_with_serialno(num);
-    if(st != NULL) {
-        passert(st->st_connection != NULL);
-        st->st_connection->extra_debugging = to_enable;
-    }
-}
-
 void recv_pcap_packet2_and_delete(u_char *user
                       , const struct pcap_pkthdr *h
                       , const u_char *bytes)
